Added in Stats Menu
Added a stats menu for player 1, 'O' to open and close the menu, 'P' to level up (For now), Numpad 7,8,9 to increase stats.  Duplicate the Canvas, and relink it up to attach it to other players, will work with the UI to make it work with movement comments to scroll through stats.